WRO-ARC-2020
We are team Taiwan Drift from Taiwan, After participating the last season of Advance Robotic challenge in WRO. We decide to share all of out robot designs to the world, Hope to contribute to the world of robotics. If you have any problems while building this robot, Feel free contact us.